+++ src/config.rs
@@ -14,11 +14,11 @@ use crate::bit_block::BitBlock;
 use crate::cache;
 use crate::cache::ReduceCache;
 use crate::primitive_array::PrimitiveArray;
-use crate::iter::{CachingBlockIter, CachingIndexIter};
+use crate::iter::{BlockIter, IndexIter};
 
 type DefaultCache = cache::FixedCache<32>;
-pub(crate) type DefaultBlockIterator<T> = CachingBlockIter<T>;
-pub(crate) type DefaultIndexIterator<T> = CachingIndexIter<T>;
+pub(crate) type DefaultBlockIterator<T> = BlockIter<T>;
+pub(crate) type DefaultIndexIterator<T> = IndexIter<T>;
 
 /// [BitSet] configuration
 /// 
@@ -58,6 +58,7 @@ pub trait Config: 'static {
     type DataBitBlock: BitBlock;
 
 // Other
+    const MAX_CAPACITY: usize;
 
     /// Cache used be [reduce()].
     /// 
@@ -66,37 +67,10 @@ pub trait Config: 'static {
 }
 
 #[inline]
-pub(crate) const fn max_addressable_index<Conf: Config>() -> usize {
+const fn max_capacity<Conf: Config>() -> usize {
     (1 << Conf::Level0BitBlock::SIZE_POT_EXPONENT)
         * (1 << Conf::Level1BitBlock::SIZE_POT_EXPONENT)
         * (1 << Conf::DataBitBlock::SIZE_POT_EXPONENT)
-}
-
-/// [SmallBitSet] configuration.
-/// 
-/// Try to keep level1 block small. Remember that [Level1BitBlock] has huge align.
-/// Try to keep [Level1MaskU64Populations] + [Level1SmallBlockIndices] size within 
-/// SIMD align.
-/// 
-/// [SmallBitSet]: crate::SmallBitSet
-/// [Level1BitBlock]: Config::Level1BitBlock
-/// [Level1MaskU64Populations]: Self::Level1MaskU64Populations
-/// [Level1SmallBlockIndices]: Self::Level1SmallBlockIndices
-pub trait SmallConfig: Config {
-    /// Small(inlined) buffer for level 1 block indices.
-    type Level1SmallBlockIndices: PrimitiveArray<
-        Item = <<Self as Config>::Level1BlockIndices as PrimitiveArray>::Item
-    >;
-    
-    /// Mask's bit-population at the start of each u64 block.
-    /// Should be [u8; Self::Mask::size()/64].
-    /// 
-    /// P.S. It is OK to use u8 even for bitblocks bigger than 256 bits,
-    /// since this array is used only when "small buffer" in use, which should
-    /// be less than 256 elements anyway.
-    ///  
-    /// P.P.S. Should be deductible from Mask, but THE RUST...  
-    type Level1MaskU64Populations: PrimitiveArray<Item=u8>;
 }
 
 /// MAX = 262_144
@@ -110,12 +84,10 @@ impl<DefaultCache: ReduceCache> Config for _64bit<DefaultCache> {
     type Level1BlockIndices = [u16; 64];
 
     type DataBitBlock = u64;
+    
+    const MAX_CAPACITY: usize = max_capacity::<Self>();
 
     type DefaultCache = DefaultCache;
-}
-impl<DefaultCache: ReduceCache> SmallConfig for _64bit<DefaultCache> {
-    type Level1SmallBlockIndices  = [u16;7];
-    type Level1MaskU64Populations = [u8;1];
 }
 
 /// MAX = 2_097_152
@@ -133,14 +105,10 @@ impl<DefaultCache: ReduceCache> Config for _128bit<DefaultCache> {
     type Level1BlockIndices = [u16; 128];
 
     type DataBitBlock = wide::u64x2;
+    
+    const MAX_CAPACITY: usize = max_capacity::<Self>();    
 
     type DefaultCache = DefaultCache;
-}
-#[cfg(feature = "simd")]
-#[cfg_attr(docsrs, doc(cfg(feature = "simd")))]
-impl<DefaultCache: ReduceCache> SmallConfig for _128bit<DefaultCache> {
-    type Level1SmallBlockIndices  = [u16;7];
-    type Level1MaskU64Populations = [u8;2];
 }
 
 /// MAX = 16_777_216 
@@ -158,12 +126,8 @@ impl<DefaultCache: ReduceCache> Config for _256bit<DefaultCache> {
     type Level1BlockIndices = [u16; 256];
 
     type DataBitBlock = wide::u64x4;
+    
+    const MAX_CAPACITY: usize = max_capacity::<Self>() - (256*256);    
 
     type DefaultCache = DefaultCache;
-}
-#[cfg(feature = "simd")]
-#[cfg_attr(docsrs, doc(cfg(feature = "simd")))]
-impl<DefaultCache: ReduceCache> SmallConfig for _256bit<DefaultCache> {
-    type Level1SmallBlockIndices  = [u16;14];
-    type Level1MaskU64Populations = [u8;4];
 }